Remove next draw page due to lottery suspension (#1063)
* chore: Show only past draws page, add claim prize button there

* chore: Remove unused components
