+++ src/views/Lottery/Lottery.tsx
@@ -1,28 +1,13 @@
 import React, { useState, useEffect } from 'react'
-import styled from 'styled-components'
-import { ButtonMenu, ButtonMenuItem } from '@pancakeswap-libs/uikit'
 import PastLotteryDataContext from 'contexts/PastLotteryDataContext'
 import { getLotteryIssueIndex } from 'utils/lotteryUtils'
-import useI18n from 'hooks/useI18n'
 import { useLottery } from 'hooks/useContract'
 import Page from 'components/layout/Page'
 import Hero from './components/Hero'
-import Divider from './components/Divider'
-import NextDrawPage from './NextDrawPage'
 import PastDrawsPage from './PastDrawsPage'
-
-const Wrapper = styled.div`
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 32px;
-`
 
 const Lottery: React.FC = () => {
   const lotteryContract = useLottery()
-  const TranslateString = useI18n()
-  const [activeIndex, setActiveIndex] = useState(0)
   const [historyData, setHistoryData] = useState([])
   const [historyError, setHistoryError] = useState(false)
   const [currentLotteryNumber, setCurrentLotteryNumber] = useState(0)
@@ -51,25 +36,14 @@ const Lottery: React.FC = () => {
     }
   }, [lotteryContract])
 
-  const handleClick = (index) => {
-    setActiveIndex(index)
-  }
-
   return (
     <>
       <Hero />
       <Page>
-        <Wrapper>
-          <ButtonMenu activeIndex={activeIndex} onItemClick={handleClick} scale="sm" variant="subtle">
-            <ButtonMenuItem>{TranslateString(716, 'Next draw')}</ButtonMenuItem>
-            <ButtonMenuItem>{TranslateString(718, 'Past draws')}</ButtonMenuItem>
-          </ButtonMenu>
-        </Wrapper>
-        <Divider />
         <PastLotteryDataContext.Provider
           value={{ historyError, historyData, mostRecentLotteryNumber, currentLotteryNumber }}
         >
-          {activeIndex === 0 ? <NextDrawPage /> : <PastDrawsPage />}
+          <PastDrawsPage />
         </PastLotteryDataContext.Provider>
       </Page>
     </>
