+++ src/views/Lottery/PastDrawsPage.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components'
 import { BaseLayout } from '@pancakeswap-libs/uikit'
 import PastLotteryRoundViewer from './components/PastLotteryRoundViewer'
 import PastDrawsHistoryCard from './components/PastDrawsHistory/PastDrawsHistoryCard'
+import ClaimPrizesCard from './components/ClaimPrizesCard'
 
 const Cards = styled(BaseLayout)`
   align-items: start;
@@ -38,15 +39,18 @@ const BunnyImageWrapper = styled.div`
 
 const PastDrawsPage: React.FC = () => {
   return (
-    <Cards>
-      <PastLotteryRoundViewer />
-      <SecondCardColumnWrapper>
-        <PastDrawsHistoryCard />
-        <BunnyImageWrapper>
-          <img src="/images/pancake-lottery-bunny.png" alt="lottery bunny" />
-        </BunnyImageWrapper>
-      </SecondCardColumnWrapper>
-    </Cards>
+    <>
+      <ClaimPrizesCard />
+      <Cards>
+        <PastLotteryRoundViewer />
+        <SecondCardColumnWrapper>
+          <PastDrawsHistoryCard />
+          <BunnyImageWrapper>
+            <img src="/images/pancake-lottery-bunny.png" alt="lottery bunny" />
+          </BunnyImageWrapper>
+        </SecondCardColumnWrapper>
+      </Cards>
+    </>
   )
 }
